[WIP] Move decode_instruction to opcode.rs
Suggested-by: Juan Bono <juan.bono@lambdaclass.com>
